requirements.txt
fastapi
uvicorn
jinja2
asyncpg
python-dotenv
requests
python-multipart